final submission that gave a score of 0.89
The evaluation metric for this competition is the Mean F1-Score. The F1 metric weights recall and precision equally, meaning a good classification model under this metric will maximize both. An average performance for both criteria is favoured over excellent performance for one criterion and poor performance for the other.